Collection refactoring & simplification
* Change PrimitiveInterface and add CollectionInterface
* Collection simplification (remove unused and non performant methods)
* Change Exception behavour
* PSR2 conformance
* Change PHPUnit by Codecption
