--- src/Collection/Validator/ObjectValidator.php
@@ -1,22 +1,0 @@
-<?php
-
-    namespace ObjectivePHP\Primitives\Collection\Validator;
-
-
-    class ObjectValidator
-    {
-
-        protected $class;
-
-        public function __construct($class = null)
-        {
-            $this->class = (string) $class;
-        }
-
-        public function __invoke($instance)
-        {
-            $expectedClass = $this->class;
-
-            return ($instance instanceof $expectedClass);
-        }
-    }
