+++ src/Exception/CollectionException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ObjectivePHP\Primitives\Exception;
+
+/**
+ * Class CollectionException
+ *
+ * @package ObjectivePHP\Primitives\Exception
+ */
+class CollectionException extends PrimitiveException
+{
+}
