+++ src/Exception/PrimitiveException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ObjectivePHP\Primitives\Exception;
+
+/**
+ * Class PrimitiveException
+ *
+ * @package src\Exception
+ */
+class PrimitiveException extends \Exception
+{
+}
