+++ src/Exception/UnderflowException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ObjectivePHP\Primitives\Exception;
+
+/**
+ * Class UnderflowException
+ *
+ * @package ObjectivePHP\Collection\Exception
+ */
+class UnderflowException extends CollectionException
+{
+}
