+++ src/Exception/UnsupportedTypeException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ObjectivePHP\Primitives\Exception;
+
+/**
+ * Class UnsupportedTypeException
+ *
+ * @package ObjectivePHP\Collection\Exception
+ */
+class UnsupportedTypeException extends CollectionException
+{
+}
