+++ tests/_support/_generated/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
