--- tests/phpunit
@@ -1,1 +1,0 @@
-../vendor/bin/phpunit
