+++ tests/unit/String/SnakeTest.php
@@ -1,18 +1,16 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: gauthier
- * Date: 04/07/2017
- * Time: 11:46
- */
 
-namespace Tests\Primitives\String;
+namespace Tests\ObjectivePHP\Primitives\String;
 
-
-use ObjectivePHP\Primitives\String\Camel;
 use ObjectivePHP\Primitives\String\Snake;
+use PHPUnit\Framework\TestCase;
 
-class SnakeTest extends \PHPUnit_Framework_TestCase
+/**
+ * Class SnakeTest
+ *
+ * @package Tests\Primitives\String
+ */
+class SnakeTest extends TestCase
 {
     /**
      * @param $snake
@@ -24,8 +22,9 @@ class SnakeTest extends \PHPUnit_Framework_TestCase
     {
         $this->assertEquals($snake, Snake::case($camel, $glue));
     }
-    
-    public function getDataForTestSnakization(){
+
+    public function getDataForTestSnakization()
+    {
         return [
             ['TestString', 'test_string'],
             ['testString', 'test_string'],
